tools/miner/cpuminer-py/requirements.txt
pyramid==1.10.4
pyblake2==1.1.2
cryptography==2.7
progressbar2==3.42